# perf(thrift): WriteStringNocopy one less func call